+++ src/main/java/com/jagrosh/jmusicbot/JMusicBot.java
@@ -37,9 +37,6 @@ import net.dv8tion.jda.api.utils.cache.CacheFlag;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import sun.misc.Signal;
-import sun.misc.SignalHandler;
- 
 
 /**
  *
@@ -66,25 +63,6 @@ public class JMusicBot
                     return;
                 default:
             }
-
-        try 
-        {
-            Signal.handle(new Signal("USR1"), new SignalHandler() 
-            {
-                public void handle(Signal sig) 
-                {
-                    System.out.println("Received message\n");
-                }
-            });
-        }
-
-        catch(IllegalArgumentException ex)
-        {
-            System.out.println( "Some aspect of the configuration is "
-            + "invalid. Illegal argument exception with signaling was thrown.");
-            System.exit(1);
-        }
-
 
         startBot();
     }
